+++ video_thumbnail_platform_interface/lib/video_thumbnail_platform_interface.dart
@@ -1,0 +1,4 @@
+library video_thumbnail_platform_interface;
+
+export 'src/image_format.dart';
+export 'src/video_thumbnail_platform.dart';