+++ video_thumbnail/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.6.0
+* Thanks for maRci002
+  - **Breaking change**: `VideoThumbnail.thumbnailFile` now returns `Future<XFile>` instead of `Future<String?>`
+  - **Breaking change**: `VideoThumbnail.thumbnailData` now returns `Future<Uint8List>` instead of `Future<Uint8List?>`
+  - migrating to use platform interface
+  - add web implementation
+
 ## 0.5.3
 * Thanks Ajb Coder for:
   -   Fix: IOException on runtime